// src/front/menterview/src/api/models/referenceModels.ts
export interface CategoryDto {
  categoryId: number;
  categoryName: string;
}

export interface DifficultyDto {
  difficultyId: number;
  difficultyName: string;
}

export interface LevelDto {
  levelId: number;
  levelName: string;
}

export interface TagDto {
  tagId: number;
  tagName: string;
}

export interface CountryDto {
  countryId: number;
  countryName: string;
}

export interface LanguageDto {
  languageId: number;
  languageName: string;
}

export interface ThemeDto {
  themeId: number;
  themeName: string;
}

export interface RoleDto {
  roleId: number;
  roleName: string;
}